requirements.txt

Pillow
gimpformats
pyora
progress
tifffile
imagecodecs
h5py